requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-role-permissions==3.2.0
gunicorn==23.0.0
idna==3.10
mercadopago==2.2.3
packaging==24.2
pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
